fix(convex): rename github-snapshot to github_snapshot
Convex doesn't allow dashes in module file paths.
